extract storage utils
